requirements.txt
opencv-contrib-python
numpy
Pillow
matplotlib
